[BREAKING CHANGE] Rename the rest of previous naming: protobufs
