+++ presenter/manager/grpc/proto/v1/manager.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package presenter.manage.grpc.proto.v1;
+package presenter.manager.grpc.proto.v1;
 
-option go_package = "github.com/teran/archived/presenter/manage/grpc/proto/v1";
+option go_package = "github.com/teran/archived/presenter/manager/grpc/proto/v1";
 
 message CreateContainerRequest {
   string name = 1;
