+++ presenter/manager/grpc/proto/v1/proto.go
@@ -1,0 +1,3 @@
+package v1
+
+//go:generate protoc -I=. --go_out=. --go_opt=paths=source_relative --go-grpc_out=. --go-grpc_opt=paths=source_relative manager.proto
